Trying to increase the accuracy of the "Connected Users"
Plan: poll all clients connected to a chatroom every x minutes or when a
new user enters. When polling, use app engine's task queue API to run a
handler in ten seconds. When polled, clients should send a message back
to the server. The relevant handler will update to the timestamp on the
connection. When the queued task finally runs (ten seconds from
polling), check timestamps. Anything older than ten seconds will be
deleted, and the connected users list will be updated on the client.